Updated SPDLog version, suppressed fmt.h warning brought from SPDLog, refactored parameter container classes in preparation for component parameter reflection

+++ GameSmithEngine/src/GameSmithEngine/Rendering/DirectX12/RenderComponents/DirectX12Shader.cpp
@@ -38,6 +38,6 @@ namespace GameSmith {
 		std::hash<std::string> hasher;
 		std::string bytecode(data, size);
 
-		m_Hash = hasher(bytecode);
+		m_Hash = (unsigned int)hasher(bytecode);
 	}
 };
